Fixed an interesting bug.  The following line does not do what many would expect:
  config.ssl_verify = @config[:ssl_verify] if @config[:ssl_verify]

The original intent of that line was to set the config.ssl_verify variable to the value of @config[:ssl_verify] if it was set.  However, if you set @config[:ssl_verify] to false, then it woun't actually set the value of config.ssl_verify because the if failed (and the default value for config.ssl_verify is true).  I changed the "if <blah>" to "unless <blah>.nil?".  I also changed the other config tests that probably would never be "false", but just in case.  :)