Fix translation issue with version 5 hiera.yaml
With the version 5 `hiera.yaml` format, we can still use older custom backends like this via `hiera3_backend`. There appears to be some magic to translate hash keys from `name` to `:name` for compatibility purposes, but it doesn't appear to recurse down into options more than one level deep, resulting in a mounts hash that looks like `"generic"=>"foo"}` instead of `:mounts=>{:generic=>"puppet"}`. This is a quick hack to just work around the problem, in lieu of a proper conversion to version 5.